implemented non-configured templates feature.
If no template configuration was found for the given template name,
mksource will try to find the $TMPLDIR/<template_name> directory and
create files for each template found in this directory.

The file name of the templates are subsituted according the substitution
rules, i.e. a file named @Target@.cpp in the template directory will
result in a <target_name>.cpp file in the output directory.